Move convertion logic out to separate file
